fix: moved camera, dancing, and lighting stuff into its older file from the main menu
